+++ apps/web/src/components/glens/bubbles/PolicyConfirmBubble.tsx
@@ -1,0 +1,116 @@
+"use client"
+import { useState } from "react"
+import { API } from "@/lib/api"
+import type { PolicyMapping } from "@/components/glens/glensTypes"
+import { SKILL_APPLY_URL, _applyBody } from "@/components/glens/glensConstants"
+
+export function PolicyConfirmBubble({
+  answer,
+  action,
+  skill,
+  draft,
+  mapping,
+  targetRuleId,
+  sessionId,
+  authFetch,
+  onResult,
+  warning,
+}: {
+  answer: string
+  action: string
+  skill: string
+  draft: Record<string, unknown>
+  mapping: PolicyMapping[]
+  targetRuleId?: string
+  sessionId: string
+  authFetch: (url: string, options?: RequestInit) => Promise<Response>
+  onResult: (text: string) => void
+  warning?: string
+}) {
+  const [status, setStatus] = useState<"pending" | "loading" | "done">("pending")
+
+  async function confirm() {
+    setStatus("loading")
+    const url = `${API}${SKILL_APPLY_URL[skill] ?? "/glens/policy/apply"}`
+    try {
+      const res = await authFetch(url, {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify(_applyBody(skill, action, draft, targetRuleId)),
+      })
+      if (!res.ok) {
+        const err = await res.json().catch(() => ({}))
+        onResult(`Failed to apply: ${err.detail ?? res.status}`)
+      } else {
+        const data = await res.json()
+        const label = data.rule_id ? `Rule "${data.rule_id}"` : data.scope ? `Budget (${data.scope})` : "Guard config"
+        onResult(`${label} ${data.action} successfully.`)
+      }
+    } catch {
+      onResult("Network error. Please try again.")
+    }
+    setStatus("done")
+  }
+
+  return (
+    <div style={{ display: "flex", justifyContent: "flex-start", marginBottom: 16, width: "100%" }}>
+      <div style={{ maxWidth: "80%", background: "var(--surface-2)", border: "1px solid var(--border)", borderRadius: "4px 14px 14px 14px", padding: "16px 20px" }}>
+        <div style={{ fontSize: 10, fontWeight: 700, color: "var(--text-muted)", textTransform: "uppercase", letterSpacing: ".06em", marginBottom: 8 }}>Policy</div>
+        <div style={{ fontSize: 14, color: "var(--text)", marginBottom: 16, lineHeight: 1.5 }}>{answer}</div>
+
+        {/* Field → Column mapping table */}
+        <table style={{ width: "100%", borderCollapse: "collapse", fontSize: 12, marginBottom: 16 }}>
+          <thead>
+            <tr style={{ borderBottom: "1px solid var(--border)" }}>
+              <th style={{ textAlign: "left", padding: "4px 8px", color: "var(--text-muted)", fontWeight: 600 }}>Field</th>
+              <th style={{ textAlign: "left", padding: "4px 8px", color: "var(--text-muted)", fontWeight: 600 }}>Value</th>
+              <th style={{ textAlign: "left", padding: "4px 8px", color: "var(--text-muted)", fontWeight: 600 }}>Column</th>
+              <th style={{ textAlign: "left", padding: "4px 8px", color: "var(--text-muted)", fontWeight: 600 }}>Description</th>
+            </tr>
+          </thead>
+          <tbody>
+            {mapping.map(m => (
+              <tr key={m.field} style={{ borderBottom: "1px solid var(--border)" }}>
+                <td style={{ padding: "6px 8px", fontFamily: "monospace", color: "var(--accent-text)" }}>{m.field}</td>
+                <td style={{ padding: "6px 8px", fontFamily: "monospace", color: "var(--text)" }}>{String(draft[m.field] ?? "—")}</td>
+                <td style={{ padding: "6px 8px", fontFamily: "monospace", color: "var(--text-2)", fontSize: 11 }}>{m.column}</td>
+                <td style={{ padding: "6px 8px", color: "var(--text-muted)" }}>{m.description}</td>
+              </tr>
+            ))}
+          </tbody>
+        </table>
+
+        {warning && (
+          <div style={{ fontSize: 12, color: "var(--warn, #f59e0b)", marginBottom: 12, padding: "8px 12px", background: "var(--warn-bg, #fef3c7)", borderRadius: 6 }}>
+            {warning}
+          </div>
+        )}
+
+        {status === "pending" && (
+          <div style={{ display: "flex", gap: 10, alignItems: "center" }}>
+            <button
+              onClick={confirm}
+              style={{
+                padding: "8px 20px", borderRadius: 8, border: "none", fontSize: 13, fontWeight: 600, cursor: "pointer",
+                background: action === "delete" ? "var(--err, #ef4444)" : "var(--accent)",
+                color: "#fff",
+              }}
+            >
+              {action === "delete" ? "Delete" : "Confirm"}
+            </button>
+            <button
+              onClick={() => onResult(`Policy ${action} cancelled.`)}
+              style={{ padding: "8px 16px", borderRadius: 8, border: "1px solid var(--border)", background: "transparent", color: "var(--text-2)", fontSize: 13, cursor: "pointer" }}
+            >
+              Cancel
+            </button>
+            {action === "delete" && (
+              <span style={{ fontSize: 11, color: "var(--err, #ef4444)" }}>This cannot be undone.</span>
+            )}
+          </div>
+        )}
+        {status === "loading" && <div style={{ fontSize: 13, color: "var(--text-muted)" }}>Applying…</div>}
+      </div>
+    </div>
+  )
+}
